refactor(services): extract callable comment-notification handler + integration tests
commentNotificationService was 945 lines with its entire notification pipeline
locked inside private methods reachable only through start() -> GraphQL
subscribe(), so none of it was directly testable. Extract the pipeline into
services/commentNotificationHandler.ts as exported, dependency-injected
functions ({ ogm, driver }); the class now only owns the subscription plumbing
(start/stop/restart) and delegates each created-comment event to
handleCommentCreatedNotification(). Behavior is unchanged.

The handler exposes the fan-out paths as standalone functions:
- handleCommentCreatedNotification (entry point: load comment, fan out)
- processDiscussionCommentNotification / processEventCommentNotification
- processCommentReplyNotification
- processFeedbackNotification / processModMention / processUserMention
- createBatchNotifications / createNotificationsForUsers / sendNotificationEmails

Email sending stays behind services/mail, which no-ops when no provider is
configured, so tests never send real mail.

Tests:
- tests/integration/commentNotifications.test.ts (5): discussion/event/reply
  notification creation against a live container, commenter exclusion, the
  opt-out reply preference, and the unknown-comment no-op.
- services/commentNotificationHandler.test.ts: the existing mock-driver unit
  tests, retargeted from the removed class methods to the extracted functions.
