+++ services/commentNotificationService.ts
@@ -1,18 +1,5 @@
 import { execute, parse, subscribe } from 'graphql';
-import {
-  createCommentNotificationEmail,
-  createEventCommentNotificationEmail,
-  createCommentReplyNotificationEmail
-} from '../customResolvers/mutations/shared/emailUtils.js';
-import { sendBatchEmails } from './mail/index.js';
-import {
-  resolveReplyNotificationRecipients,
-  type NotificationRecipient,
-} from './commentNotificationRecipients.js';
-import { notifyFeedback } from '../hooks/feedbackNotificationHook.js';
-import { notifyModMentions } from '../hooks/modMentionNotificationHook.js';
-import { notifyCommentMentions } from '../hooks/userMentionNotificationHook.js';
-import { appendNotificationFooter } from '../utils/notificationFooter.js';
+import { handleCommentCreatedNotification } from './commentNotificationHandler.js';
 
 type AsyncIterableIterator<T> = AsyncIterable<T> & AsyncIterator<T>;
 
@@ -110,12 +97,14 @@ export class CommentNotificationService {
 
         const commentBasicInfo = result.data.commentCreated.createdComment;
         const commentId = commentBasicInfo.id;
-        
+
         console.log('Processing notification for newly created comment:', commentId);
 
         try {
-          // Fetch the full comment details
-          await this.processCommentNotification(commentId);
+          await handleCommentCreatedNotification(
+            { ogm: this.ogm, driver: this.driver },
+            commentId
+          );
         } catch (error) {
           console.error('Error processing comment notification:', error);
           // Continue processing other events even if one fails
@@ -129,806 +118,6 @@ export class CommentNotificationService {
         console.log('Restarting comment notification service in 5 seconds...');
         setTimeout(() => this.start(), 5000);
       }
-    }
-  }
-
-  /**
-   * Process a notification for a newly created comment
-   */
-  private async processCommentNotification(commentId: string) {
-    console.log('=== DEBUG: Starting processCommentNotification for comment:', commentId);
-    
-    // Get the OGM models
-    const CommentModel = this.ogm.model('Comment');
-    const UserModel = this.ogm.model('User');
-    const DiscussionModel = this.ogm.model('Discussion');
-
-    // Fetch the full comment details with a more comprehensive selection
-    const fullComments = await CommentModel.find({
-      where: { id: commentId },
-      selectionSet: `{
-        id
-        text
-        isRootComment
-        isFeedbackComment
-        Channel {
-          uniqueName
-        }
-        CommentAuthor {
-          ... on User {
-            __typename
-            username
-            displayName
-          }
-          ... on ModerationProfile {
-            __typename
-            displayName
-            User {
-              username
-            }
-          }
-        }
-        DiscussionChannel {
-          id
-          discussionId
-          channelUniqueName
-          Channel {
-            uniqueName
-            displayName
-          }
-          Discussion {
-            id
-            title
-            Author {
-              username
-            }
-          }
-          SubscribedToNotifications {
-            username
-          }
-        }
-        Event {
-          id
-          title
-          Poster {
-            username
-          }
-          EventChannels {
-            channelUniqueName
-            Channel {
-              uniqueName
-            }
-          }
-          SubscribedToNotifications {
-            username
-          }
-        }
-        GivesFeedbackOnComment {
-          id
-          CommentAuthor {
-            ... on User {
-              username
-            }
-            ... on ModerationProfile {
-              User {
-                username
-              }
-            }
-          }
-        }
-        GivesFeedbackOnDiscussion {
-          id
-          Author {
-            username
-          }
-        }
-        GivesFeedbackOnEvent {
-          id
-          Poster {
-            username
-          }
-        }
-        ParentComment {
-          id
-          text
-          CommentAuthor {
-            ... on User {
-              __typename
-              username
-              notifyOnReplyToCommentByDefault
-              Email {
-                address
-              }
-            }
-            ... on ModerationProfile {
-              __typename
-              displayName
-              User {
-                username
-                notifyOnReplyToCommentByDefault
-                Email {
-                  address
-                }
-              }
-            }
-          }
-          SubscribedToNotifications {
-            username
-            Email {
-              address
-            }
-          }
-        }
-      }`
-    });
-
-    if (!fullComments || !fullComments.length) {
-      console.error('=== DEBUG ERROR: Could not find comment details for ID:', commentId);
-      return;
-    }
-
-    const fullComment = fullComments[0];
-    console.log('=== DEBUG: Found comment details for ID:', commentId);
-    console.log('=== DEBUG: Comment structure:', {
-      hasDiscussionChannel: !!fullComment.DiscussionChannel,
-      hasEvent: !!fullComment.Event,
-      hasParentComment: !!fullComment.ParentComment,
-      discussionChannelId: fullComment.DiscussionChannel?.id,
-      eventId: fullComment.Event?.id,
-      parentCommentId: fullComment.ParentComment?.id,
-      channelName: fullComment.Channel?.uniqueName
-    });
-
-    // Get the commenter info
-    const commenterUsername =
-      fullComment.CommentAuthor?.username ||
-      fullComment.CommentAuthor?.User?.username ||
-      'Someone';
-
-    const commenterDisplayName =
-      fullComment.CommentAuthor?.displayName ||
-      commenterUsername;
-
-    const authorLabel = `[@${commenterDisplayName}](/u/${commenterUsername})`;
-
-    console.log('Processing notification for comment type:');
-    console.log('- Has DiscussionChannel:', !!fullComment.DiscussionChannel);
-    console.log('- Has Event:', !!fullComment.Event);
-    console.log('- Has ParentComment:', !!fullComment.ParentComment);
-    console.log('- Is feedback comment:', !!fullComment.isFeedbackComment);
-    console.log('- Commenter username:', commenterUsername);
-
-    const channelUniqueName = fullComment.Channel?.uniqueName ||
-      fullComment.DiscussionChannel?.channelUniqueName ||
-      fullComment.Event?.EventChannels?.[0]?.channelUniqueName;
-
-    // FEEDBACK NOTIFICATION
-    // When someone gives feedback on content, notify the content author
-    if (fullComment.isFeedbackComment) {
-      console.log('=== DEBUG: Processing FEEDBACK comment notification');
-      await this.processFeedbackNotification(fullComment, commentId, commenterUsername, commenterDisplayName, channelUniqueName);
-    }
-
-    // MOD MENTION NOTIFICATIONS
-    // Check for /m/modProfileName mentions and notify those moderators
-    await this.processModMentionNotifications(fullComment, commentId, commenterUsername, authorLabel, channelUniqueName);
-
-    // USER MENTION NOTIFICATIONS
-    // Check for @username mentions and notify those users
-    await this.processUserMentionNotifications(fullComment, commentId, commenterUsername, commenterDisplayName, channelUniqueName);
-
-    // COMMENT REPLY NOTIFICATION
-    // Check for replies first, as a reply can also have a DiscussionChannel or Event
-    if (fullComment.ParentComment) {
-      console.log('This is a REPLY to another comment');
-      await this.processCommentReplyNotification(
-        fullComment,
-        commentId,
-        commenterUsername,
-        CommentModel,
-        UserModel
-      );
-    }
-    // DISCUSSION COMMENT NOTIFICATION
-    else if (fullComment.DiscussionChannel) {
-      console.log('=== DEBUG: This is a comment on a DISCUSSION');
-      console.log('=== DEBUG: DiscussionChannel details:', {
-        id: fullComment.DiscussionChannel.id,
-        discussionId: fullComment.DiscussionChannel.discussionId,
-        channelUniqueName: fullComment.DiscussionChannel.channelUniqueName,
-        subscribedUsersCount: fullComment.DiscussionChannel.SubscribedToNotifications?.length || 0,
-        subscribedUsers: fullComment.DiscussionChannel.SubscribedToNotifications?.map((u: any) => u.username) || []
-      });
-      await this.processDiscussionCommentNotification(
-        fullComment,
-        commentId,
-        commenterUsername,
-        DiscussionModel,
-        UserModel
-      );
-    }
-    // EVENT COMMENT NOTIFICATION
-    else if (fullComment.Event) {
-      console.log('This is a comment on an EVENT');
-      await this.processEventCommentNotification(
-        fullComment,
-        commentId,
-        commenterUsername,
-        UserModel
-      );
-    }
-  }
-
-  /**
-   * Create batch notifications and send emails for subscribed users using Cypher
-   */
-  private async createBatchNotifications(
-    driver: any,
-    notificationText: string,
-    commenterUsername: string,
-    entityType: 'DiscussionChannel' | 'Event' | 'Comment',
-    entityId: string,
-    emailContent?: { subject: string; plainText: string; html: string },
-    notificationType: string = 'reply'
-  ) {
-    console.log('=== DEBUG: Starting createBatchNotifications');
-    console.log('=== DEBUG: Parameters:', {
-      entityType,
-      entityId,
-      commenterUsername,
-      notificationTextLength: notificationText.length
-    });
-    
-    const session = driver.session();
-    
-    try {
-      // First, let's check what subscriptions exist using OGM to get email addresses
-      console.log('=== DEBUG: Fetching subscribed users with emails using OGM');
-      
-      // Get the entity model based on entityType
-      const EntityModel = this.ogm.model(entityType);
-      
-      // Fetch the entity with its subscribed users and their emails
-      const entityResults = await EntityModel.find({
-        where: { id: entityId },
-        selectionSet: `{
-          id
-          SubscribedToNotifications {
-            username
-            Email {
-              address
-            }
-          }
-        }`
-      });
-      
-      if (!entityResults || !entityResults.length) {
-        console.error('=== DEBUG ERROR: Entity not found:', { entityType, entityId });
-        return 0;
-      }
-      
-      const entity = entityResults[0];
-      const subscribedUsersData = entity.SubscribedToNotifications?.map((user: any) => ({
-        username: user.username,
-        email: user.Email?.address || null
-      })) || [];
-      
-      console.log('=== DEBUG: Found subscribed users with emails:', subscribedUsersData);
-      console.log('=== DEBUG: Total subscribed users count:', subscribedUsersData.length);
-      
-      // Filter out the commenter
-      const usersToNotify = subscribedUsersData.filter((userData: any) => userData.username !== commenterUsername);
-      console.log('=== DEBUG: Users to notify (excluding commenter):', usersToNotify.map((u: any) => u.username));
-      
-      if (usersToNotify.length === 0) {
-        console.log('=== DEBUG: No users to notify after filtering out commenter');
-        return 0;
-      }
-      
-      // Send batch emails if email content is provided
-      if (emailContent) {
-        await this.sendBatchEmails(usersToNotify, emailContent);
-      }
-      
-      // Create notifications using Cypher (this part stays the same)
-      const cypherQuery = `
-        MATCH (entity:${entityType} {id: $entityId})
-        MATCH (entity)<-[:SUBSCRIBED_TO_NOTIFICATIONS]-(user:User)
-        WHERE user.username <> $commenterUsername
-        CREATE (notification:Notification {
-          id: randomUUID(),
-          createdAt: datetime(),
-          read: false,
-          text: $notificationText,
-          notificationType: $notificationType
-        })
-        CREATE (user)-[:HAS_NOTIFICATION]->(notification)
-        RETURN count(notification) as notificationsCreated, collect(user.username) as notifiedUsers
-      `;
-
-      console.log('=== DEBUG: Executing notification creation query:', cypherQuery);
-      const result = await session.run(cypherQuery, {
-        entityId,
-        commenterUsername,
-        notificationText,
-        notificationType
-      });
-
-      const notificationsCreated = result.records[0]?.get('notificationsCreated')?.toNumber() || 0;
-      const notifiedUsers = result.records[0]?.get('notifiedUsers') || [];
-      
-      console.log('=== DEBUG: Notification creation results:', {
-        notificationsCreated,
-        notifiedUsers,
-        entityType,
-        entityId
-      });
-      
-      if (notificationsCreated === 0) {
-        console.log('=== DEBUG WARNING: No notifications were created');
-      }
-      
-      return notificationsCreated;
-    } catch (error) {
-      console.error('=== DEBUG ERROR: Error in createBatchNotifications:', error);
-      throw error;
-    } finally {
-      session.close();
-    }
-  }
-
-  /**
-   * Process notification for a comment on a discussion
-   */
-  private async processDiscussionCommentNotification(
-    fullComment: any,
-    commentId: string,
-    commenterUsername: string,
-    DiscussionModel: any,
-    UserModel: any
-  ) {
-    console.log('=== DEBUG: Starting processDiscussionCommentNotification');
-    console.log('=== DEBUG: Commenter username:', commenterUsername);
-
-    const discussionChannel = fullComment.DiscussionChannel;
-    if (!discussionChannel) {
-      console.error('=== DEBUG ERROR: No discussion channel found');
-      return;
-    }
-    
-    console.log('=== DEBUG: DiscussionChannel ID:', discussionChannel.id);
-
-    const discussion = discussionChannel.Discussion;
-    if (!discussion) {
-      console.error('=== DEBUG ERROR: Discussion not found in DiscussionChannel');
-      return;
-    }
-    
-    console.log('=== DEBUG: Discussion details:', {
-      id: discussion.id,
-      title: discussion.title,
-      authorUsername: discussion.Author?.username
-    });
-
-    const channelName = fullComment.Channel?.uniqueName;
-    const discussionUrl = `${process.env.FRONTEND_URL}/forums/${channelName}/discussions/${discussion.id}`;
-
-    // Create notification text for in-app notification with unsubscribe footer
-    const baseNotificationText = `${commenterUsername} commented on the discussion [${discussion.title}](${discussionUrl}/comments/${commentId})`;
-    const notificationText = appendNotificationFooter(baseNotificationText, {
-      contentType: 'discussion',
-      contentUrl: discussionUrl,
-    });
-
-    // Create email content for discussion notification
-    const emailContent = createCommentNotificationEmail(
-      fullComment.text,
-      discussion.title,
-      commenterUsername,
-      channelName || '',
-      discussion.id,
-      commentId
-    );
-
-    console.log('=== DEBUG: Creating batch notifications for discussion comment:', discussion.title);
-    console.log('=== DEBUG: Notification text:', notificationText);
-
-    // Use batch Cypher query to create notifications for all subscribed users
-    if (this.driver) {
-      console.log('=== DEBUG: Calling createBatchNotifications with:', {
-        entityType: 'DiscussionChannel',
-        entityId: discussionChannel.id,
-        commenterUsername,
-        driver: !!this.driver,
-        hasEmailContent: !!emailContent
-      });
-      const notificationsCreated = await this.createBatchNotifications(
-        this.driver,
-        notificationText,
-        commenterUsername,
-        'DiscussionChannel',
-        discussionChannel.id,
-        emailContent,
-        'reply'
-      );
-      console.log('=== DEBUG: Batch notifications result:', notificationsCreated);
-    } else {
-      console.error('=== DEBUG ERROR: Driver not available for batch notifications');
-    }
-  }
-
-  /**
-   * Process notification for a comment on an event
-   */
-  private async processEventCommentNotification(
-    fullComment: any,
-    commentId: string,
-    commenterUsername: string,
-    UserModel: any
-  ) {
-    console.log('Processing comment on event');
-
-    const event = fullComment.Event;
-    if (!event) {
-      console.log('Event not found');
-      return;
-    }
-
-    const channelName = fullComment.Channel?.uniqueName;
-    if (!channelName) {
-      console.log('No channel found for event');
-      return;
-    }
-
-    const eventUrl = `${process.env.FRONTEND_URL}/forums/${channelName}/events/${event.id}`;
-
-    // Create notification text for in-app notification with unsubscribe footer
-    const baseNotificationText = `${commenterUsername} commented on the event [${event.title}](${eventUrl}/comments/${commentId})`;
-    const notificationText = appendNotificationFooter(baseNotificationText, {
-      contentType: 'event',
-      contentUrl: eventUrl,
-    });
-
-    // Create email content for event notification
-    const emailContent = createEventCommentNotificationEmail(
-      fullComment.text,
-      event.title,
-      commenterUsername,
-      channelName,
-      event.id,
-      commentId
-    );
-
-    console.log('=== DEBUG: Creating batch notifications for event comment:', event.title);
-
-    // Use batch Cypher query to create notifications for all subscribed users
-    if (this.driver) {
-      await this.createBatchNotifications(
-        this.driver,
-        notificationText,
-        commenterUsername,
-        'Event',
-        event.id,
-        emailContent,
-        'reply'
-      );
-    } else {
-      console.error('=== DEBUG ERROR: Driver not available for batch notifications');
-    }
-  }
-
-  /**
-   * Process notification for a reply to a comment
-   */
-  private async processCommentReplyNotification(
-    fullComment: any,
-    commentId: string,
-    commenterUsername: string,
-    CommentModel: any,
-    UserModel: any
-  ) {
-    console.log('Processing reply to comment');
-
-    const parentComment = fullComment.ParentComment;
-    if (!parentComment) {
-      console.log('Parent comment not found');
-      return;
-    }
-
-    const parentCommentId = parentComment.id;
-    console.log('Parent comment ID:', parentCommentId);
-
-    // Determine content title and URL based on parent comment's context
-    let contentTitle: string;
-    let contentUrl: string;
-    let commentPermalinkUrl: string;
-    let channelName: string | undefined;
-
-    // Check if the reply is on a discussion or event (from the current comment's context)
-    if (fullComment.DiscussionChannel) {
-      const discussion = fullComment.DiscussionChannel.Discussion;
-      contentTitle = discussion?.title || 'a discussion';
-      channelName = fullComment.Channel?.uniqueName;
-      contentUrl = `${process.env.FRONTEND_URL}/forums/${channelName}/discussions/${discussion?.id}`;
-      commentPermalinkUrl = `${contentUrl}/comments/${parentCommentId}`;
-    } else if (fullComment.Event) {
-      const event = fullComment.Event;
-      contentTitle = event?.title || 'an event';
-      channelName = fullComment.Channel?.uniqueName;
-      contentUrl = `${process.env.FRONTEND_URL}/forums/${channelName}/events/${event?.id}`;
-      commentPermalinkUrl = `${contentUrl}/comments/${parentCommentId}`;
-    } else {
-      console.log('No content reference found for comment reply');
-      return;
-    }
-
-    // Create notification text for in-app notification with unsubscribe footer
-    const baseNotificationText = `${commenterUsername} replied to your comment on [${contentTitle}](${commentPermalinkUrl})`;
-    const notificationText = appendNotificationFooter(baseNotificationText, {
-      contentType: 'comment',
-      contentUrl: commentPermalinkUrl,
-    });
-
-    // Create email content for reply notification
-    const emailContent = createCommentReplyNotificationEmail(
-      fullComment.text,
-      contentTitle,
-      commenterUsername,
-      commentPermalinkUrl
-    );
-
-    console.log('=== DEBUG: Creating batch notifications for comment reply on:', contentTitle);
-
-    const usersToNotify = resolveReplyNotificationRecipients({
-      commenterUsername,
-      parentCommentAuthor: parentComment.CommentAuthor,
-      subscribedUsers: parentComment.SubscribedToNotifications,
-    });
-
-    console.log('=== DEBUG: Reply notification recipients:', usersToNotify);
-
-    if (usersToNotify.length === 0) {
-      console.log('=== DEBUG: No reply recipients after applying author preference and subscriptions');
-      return;
-    }
-
-    if (this.driver) {
-      await this.createNotificationsForUsers(
-        this.driver,
-        usersToNotify,
-        notificationText,
-        emailContent,
-        'reply'
-      );
-    } else {
-      console.error('=== DEBUG ERROR: Driver not available for batch notifications');
-    }
-  }
-
-  /**
-   * Send batch emails using SendGrid
-   */
-  private async sendBatchEmails(
-    usersToNotify: NotificationRecipient[],
-    emailContent: { subject: string; plainText: string; html: string }
-  ) {
-    try {
-      // Filter users who have email addresses
-      const usersWithEmails = usersToNotify.filter(user => user.email);
-      
-      console.log('=== DEBUG: Users with emails to notify:', usersWithEmails.length);
-      
-      if (usersWithEmails.length === 0) {
-        console.log('=== DEBUG: No users with email addresses to notify');
-        return;
-      }
-      
-      // Prepare batch email data
-      const emailsToSend = usersWithEmails.map(user => ({
-        to: user.email!,
-        subject: emailContent.subject,
-        text: emailContent.plainText,
-        html: emailContent.html
-      }));
-      
-      console.log('=== DEBUG: Sending batch emails to:', emailsToSend.map(email => email.to));
-      
-      // Send batch emails
-      const emailSent = await sendBatchEmails(emailsToSend);
-
-      if (!emailSent) {
-        console.log('=== DEBUG: Batch email send skipped or failed');
-        return;
-      }
-      
-      console.log('=== DEBUG: Successfully sent', emailsToSend.length, 'batch emails');
-      
-    } catch (error) {
-      console.error('=== DEBUG ERROR: Failed to send batch emails:', error);
-      // Don't throw - continue with in-app notifications even if emails fail
-    }
-  }
-
-  private async createNotificationsForUsers(
-    driver: any,
-    usersToNotify: NotificationRecipient[],
-    notificationText: string,
-    emailContent?: { subject: string; plainText: string; html: string },
-    notificationType: string = 'reply'
-  ) {
-    const session = driver.session();
-
-    try {
-      if (emailContent) {
-        await this.sendBatchEmails(usersToNotify, emailContent);
-      }
-
-      const usernames = usersToNotify.map((user) => user.username);
-      console.log('=== DEBUG: Creating direct notifications for users:', usernames);
-
-      const result = await session.run(
-        `
-        UNWIND $usernames AS username
-        MATCH (user:User {username: username})
-        CREATE (notification:Notification {
-          id: randomUUID(),
-          createdAt: datetime(),
-          read: false,
-          text: $notificationText,
-          notificationType: $notificationType
-        })
-        CREATE (user)-[:HAS_NOTIFICATION]->(notification)
-        RETURN count(notification) as notificationsCreated, collect(user.username) as notifiedUsers
-        `,
-        {
-          usernames,
-          notificationText,
-          notificationType,
-        }
-      );
-
-      const notificationsCreated = result.records[0]?.get('notificationsCreated')?.toNumber() || 0;
-      const notifiedUsers = result.records[0]?.get('notifiedUsers') || [];
-
-      console.log('=== DEBUG: Direct notification creation results:', {
-        notificationsCreated,
-        notifiedUsers,
-      });
-
-      return notificationsCreated;
-    } finally {
-      session.close();
-    }
-  }
-
-  /**
-   * Process feedback notification - notify the content author that they received feedback
-   */
-  private async processFeedbackNotification(
-    fullComment: any,
-    commentId: string,
-    authorUsername: string,
-    authorDisplayName: string,
-    channelUniqueName: string | null
-  ) {
-    try {
-      // Determine who should receive the feedback notification
-      let targetAuthorUsername: string | null = null;
-      let targetType: 'comment' | 'discussion' | 'event' = 'comment';
-      let targetId: string | null = null;
-      let discussionId: string | null = null;
-      let eventId: string | null = null;
-
-      if (fullComment.GivesFeedbackOnComment) {
-        const feedbackTarget = fullComment.GivesFeedbackOnComment;
-        targetAuthorUsername =
-          feedbackTarget.CommentAuthor?.username ||
-          feedbackTarget.CommentAuthor?.User?.username;
-        targetType = 'comment';
-        targetId = feedbackTarget.id;
-        discussionId = fullComment.DiscussionChannel?.discussionId;
-        eventId = fullComment.Event?.id;
-      } else if (fullComment.GivesFeedbackOnDiscussion) {
-        const feedbackTarget = fullComment.GivesFeedbackOnDiscussion;
-        targetAuthorUsername = feedbackTarget.Author?.username;
-        targetType = 'discussion';
-        targetId = feedbackTarget.id;
-        discussionId = feedbackTarget.id;
-      } else if (fullComment.GivesFeedbackOnEvent) {
-        const feedbackTarget = fullComment.GivesFeedbackOnEvent;
-        targetAuthorUsername = feedbackTarget.Poster?.username;
-        targetType = 'event';
-        targetId = feedbackTarget.id;
-        eventId = feedbackTarget.id;
-      }
-
-      if (!targetAuthorUsername || !targetId) {
-        console.log('=== DEBUG: No feedback target found');
-        return;
-      }
-
-      console.log('=== DEBUG: Sending feedback notification to:', targetAuthorUsername);
-
-      await notifyFeedback({
-        context: { ogm: this.ogm, driver: this.driver },
-        feedbackContext: {
-          feedbackCommentId: commentId,
-          authorUsername,
-          authorDisplayName,
-          targetType,
-          targetId,
-          channelUniqueName: channelUniqueName || undefined,
-          discussionId: discussionId || undefined,
-          eventId: eventId || undefined,
-        },
-        targetAuthorUsername,
-      });
-    } catch (error) {
-      console.error('Error processing feedback notification:', error);
-    }
-  }
-
-  /**
-   * Process mod mention notifications - notify moderators mentioned with /m/modProfileName
-   */
-  private async processModMentionNotifications(
-    fullComment: any,
-    commentId: string,
-    authorUsername: string,
-    authorLabel: string,
-    channelUniqueName: string | null
-  ) {
-    try {
-      const discussionChannel = fullComment.DiscussionChannel;
-      const event = fullComment.Event;
-
-      await notifyModMentions({
-        context: { ogm: this.ogm, driver: this.driver },
-        mentionContext: {
-          type: 'comment',
-          authorUsername,
-          authorLabel,
-          commentId,
-          discussionId: discussionChannel?.discussionId,
-          discussionTitle: discussionChannel?.Discussion?.title,
-          eventId: event?.id,
-          eventTitle: event?.title,
-          channelUniqueName: channelUniqueName || undefined,
-        },
-        previousText: null, // New comment, no previous text
-        nextText: fullComment.text,
-      });
-    } catch (error) {
-      console.error('Error processing mod mention notifications:', error);
-    }
-  }
-
-  /**
-   * Process user mention notifications - notify users mentioned with @username
-   */
-  private async processUserMentionNotifications(
-    fullComment: any,
-    commentId: string,
-    _authorUsername: string,
-    _authorDisplayName: string,
-    _channelUniqueName: string | null
-  ) {
-    try {
-      // Pass the full comment which matches CommentSnapshot structure
-      await notifyCommentMentions({
-        context: { ogm: this.ogm, driver: this.driver },
-        comment: {
-          id: commentId,
-          text: fullComment.text,
-          CommentAuthor: fullComment.CommentAuthor,
-          DiscussionChannel: fullComment.DiscussionChannel,
-          Event: fullComment.Event,
-        },
-        previousText: null, // New comment, no previous text
-        nextText: fullComment.text,
-      });
-    } catch (error) {
-      console.error('Error processing user mention notifications:', error);
     }
   }
